==> Figure1/06_make_pretty_heatmap_monoNucs_occu.Rmd <==
---
title: "make pretty heatmap mononucs"
author: "Uwe Schwartz"
date: "02/12/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(warning = FALSE)
USER<-Sys.info()["user"]
work.path<- paste0("/home/",USER,"/nucMACC_Paper/R001_nucMacc/manuscript_figures/Fig1/heatmaps/monoNucs_occu/")
knitr::opts_knit$set(root.dir=work.path)
#setwd(work.path)
```

## Dependencies & Libraries

#### Load packages
```{r libraries, message=FALSE}
library(pheatmap)
library(RColorBrewer)
library(ggpubr)
```

## Import data
```{r import}
df.rpkm<-read.delim(file="../../data/expression/RPKM_protCod_genes.txt")

head(df.rpkm)

### load gene sorting of heatmap
genes.list.H4<-read.delim("sortedRegions_Heatmap_H4.txt")
head(genes.list.H4)

genes.heat.sorted.na.H4<-genes.list.H4[match(as.character(df.rpkm$refSeq), 
                                       as.character(genes.list.H4$name), nomatch = ),]
#remove NAs
table(is.na(match(as.character(df.rpkm$refSeq), as.character(genes.list.H4$name))))

genes.heat.sorted.H4<-genes.heat.sorted.na.H4[!is.na(match(as.character(df.rpkm$refSeq),
                                    as.character(genes.list.H4$name))),]
head(genes.heat.sorted.H4)

## load heatmap
heat.val.H4<-read.delim("values_Heatmap_H4.txt", skip=2)
row.names(heat.val.H4)<-as.character(genes.list.H4$name)
heat.val.sort.H4<-heat.val.H4[as.character(genes.heat.sorted.H4$name),]
head(heat.val.sort.H4[,1:10])
```
## Analysis

```{r Analysis}
# get samples from heatmap
# mono.nuc occupancy
H4.mono.occ.col<-grep("pooled_monoNucs_profile",
                      colnames(heat.val.sort.H4), value = T)

genes.split<-split(df.rpkm$refSeq,f = df.rpkm$expr.strength)

profile.list<-lapply(genes.split, function(x) 
    apply(heat.val.sort.H4[which(rownames(heat.val.sort.H4) %in% x),
                        H4.mono.occ.col[75:250]],
                                      2,function(x) mean(x,na.rm = T)))

pdf("profile_H4_monoNucs_occ.pdf", width = 4, height=3)
    plot(seq(-750,1000,10), profile.list[["silent"]], type="l", 
         xlab="distance from TSS", ylab = "mono-nucleosome occupancy",
         lwd=2, main="H4 mono-nucs", col="#5CB85B", ylim=c(0,270))
    lines(seq(-750,1000,10), profile.list[["low"]], col="#FEA22B",
         lwd=2)
    lines(seq(-750,1000,10), profile.list[["mid"]], col="#357EBD",
         lwd=2)
    lines(seq(-750,1000,10), profile.list[["high"]], col="#D43F39",
         lwd=2)
dev.off()

# heatmap
colNucs<-colorRampPalette(c(rep("#FFFFFF",1),"#D95F02"))(100)

heat.mono.H4<-heat.val.sort.H4[,H4.mono.occ.col[75:250]]
colnames(heat.mono.H4)<-c("-750", rep(".",75),"TSS", rep(".",98),"+ 1 kb")

# heatmap
pheatmap(heat.mono.H4[1:2500,], scale="none",
         breaks=seq(0,200,length.out = 101),
         show_colnames = T, show_rownames = F,
         color = colNucs,
         cluster_rows = F, cluster_cols = F,
         na_col = "#FFFFFF")
dev.off()

png("H4_mono-nuc_occ_heatmap_white.png",res = 700, width = 2000, height = 2800)
pheatmap(heat.mono.H4, scale="none",
         breaks=seq(0,200,length.out = 101),
         show_colnames = T, show_rownames = F,
         color = colNucs,
         cluster_rows = F, cluster_cols = F,
         na_col = "#FFFFFF")
dev.off()
```
## H3

```{r H3}
# load gene sorting of heatmap
genes.list.H3<-read.delim("sortedRegions_Heatmap_H3.txt")
head(genes.list.H3)

genes.heat.sorted.na.H3<-genes.list.H3[match(as.character(df.rpkm$refSeq), 
                                       as.character(genes.list.H3$name), nomatch = ),]
# remove NAs
table(is.na(match(as.character(df.rpkm$refSeq), as.character(genes.list.H3$name))))

genes.heat.sorted.H3<-genes.heat.sorted.na.H3[!is.na(match(as.character(df.rpkm$refSeq),
                                    as.character(genes.list.H3$name))),]
head(genes.heat.sorted.H3)

# load heatmap
heat.val.H3<-read.delim("values_Heatmap_H3.txt", skip=2)
row.names(heat.val.H3)<-as.character(genes.list.H3$name)

heat.val.sort.H3<-heat.val.H3[as.character(genes.heat.sorted.H3$name),]
head(heat.val.sort.H3[,1:10])

# get samples from heatmap
# mono.nuc occupancy
H3.mono.occ.col<-grep("pooled_monoNucs_profile",
                      colnames(heat.val.sort.H3), value = T)

genes.split<-split(df.rpkm$refSeq,f = df.rpkm$expr.strength)

profile.list<-lapply(genes.split, function(x) 
    apply(heat.val.sort.H3[which(rownames(heat.val.sort.H3) %in% x),
                        H3.mono.occ.col[75:250]],
                                      2,function(x) mean(x,na.rm = T)))

pdf("profile_H3_monoNucs_occ.pdf", width = 4, height=3)
    plot(seq(-750,1000,10), profile.list[["silent"]], type="l", 
         xlab="distance from TSS", ylab = "mono-nucleosome occupancy",
         lwd=2, main="H3 mono-nucs", col="#5CB85B", ylim=c(0,270))
    lines(seq(-750,1000,10), profile.list[["low"]], col="#FEA22B",
         lwd=2)
    lines(seq(-750,1000,10), profile.list[["mid"]], col="#357EBD",
         lwd=2)
    lines(seq(-750,1000,10), profile.list[["high"]], col="#D43F39",
         lwd=2)
dev.off()

# heatmap
colNucs<-colorRampPalette(c(rep("#FFFFFF",1),"#D95F02"))(100)
heat.mono.H3<-heat.val.sort.H3[,H3.mono.occ.col[75:250]]
colnames(heat.mono.H3)<-c("-750", rep(".",75),"TSS", rep(".",98),"+ 1 kb")

# heatmap
pheatmap(heat.mono.H3[1:2500,], scale="none",
         breaks=seq(0,200,length.out = 101),
         show_colnames = T, show_rownames = F,
         color = colNucs,
         cluster_rows = F, cluster_cols = F,
         na_col = "#FFFFFF")
dev.off()

png("H3_mono-nuc_occ_heatmap_white.png",res = 700, width = 2000, height = 2800)
pheatmap(heat.mono.H3, scale="none",
         breaks=seq(0,200,length.out = 101),
         show_colnames = T, show_rownames = F,
         color = colNucs,
         cluster_rows = F, cluster_cols = F,
         na_col = "#FFFFFF")
dev.off()
```

## Session info
```{r info}
sessionInfo()
```
